covelo/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Brotli==1.0.9
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.3.0
Django==4.1.7
djangorestframework==3.14.0
drf-yasg==1.21.5
exponent-server-sdk==2.0.0
gunicorn==20.1.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
mysqlclient==2.1.1
packaging==23.1
Pillow==9.5.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pytz==2023.3
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.15
whitenoise==6.4.0
